--- wrench/components/cataloger/__init__.py
@@ -1,4 +1,0 @@
-from .base import Cataloger
-from .sddi import SDDICataloger
-
-__all__ = ["Cataloger", "SDDICataloger"]
