--- wrench/components/grouper/__init__.py
@@ -1,4 +1,0 @@
-from .base import Grouper
-from .teleclass import TELEClassGrouper
-
-__all__ = ["Grouper", "TELEClassGrouper"]
