--- wrench/components/harvester/__init__.py
@@ -1,4 +1,0 @@
-from .base import Harvester
-from .sensorthings import SensorThingsHarvesterComponent
-
-__all__ = ["SensorThingsHarvesterComponent", "Harvester"]
