--- wrench/components/metadata_builder/__init__.py
@@ -1,4 +1,0 @@
-from .base import MetadataBuilder
-from .sensorthings import SensorThingsMetadataBuilder
-
-__all__ = ["MetadataBuilder", "SensorThingsMetadataBuilder"]
